+++ app/src/main/java/com/stopsmoke/kekkek/core/data/mapper/CommentMapper.kt
@@ -9,8 +9,7 @@ import com.stopsmoke.kekkek.core.firestore.model.CommentEntity
 import com.stopsmoke.kekkek.core.firestore.model.CommentParentEntity
 import com.stopsmoke.kekkek.core.firestore.model.ReplyEntity
 import com.stopsmoke.kekkek.core.firestore.model.WrittenEntity
-import com.stopsmoke.kekkek.presentation.toPostCategory
-import com.stopsmoke.kekkek.presentation.toRequestString
+import com.stopsmoke.kekkek.presentation.mapper.toPostCategory
 
 internal fun CommentEntity.asExternalModel(
     earliestReply: List<Reply>,
