+++ app/src/main/java/com/stopsmoke/kekkek/presentation/Utils.kt
@@ -135,16 +135,6 @@ internal fun<T> Bundle.getParcelableAndroidVersionSupport(key: String, clazz: Cl
         getParcelable(key)
     }
 
-internal fun PostCategory.getResourceString(context: Context) = when(this) {
-    PostCategory.NOTICE -> context.getString(R.string.community_category_notice)
-    PostCategory.QUIT_SMOKING_AIDS_REVIEWS -> context.getString(R.string.community_category_quit_smoking_aids_reviews)
-    PostCategory.SUCCESS_STORIES -> context.getString(R.string.community_category_success_stories)
-    PostCategory.GENERAL_DISCUSSION -> context.getString(R.string.community_category_general_discussion)
-    PostCategory.RESOLUTIONS -> context.getString(R.string.community_category_resolutions)
-    PostCategory.UNKNOWN -> context.getString(R.string.community_category_unknown)
-    PostCategory.ALL -> context.getString(R.string.community_category_all)
-}
-
 fun isNetworkAvailable(context: Context): Boolean {
     val connectivityManager =
         context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
@@ -184,55 +174,6 @@ internal fun<T> NavController.putNavigationResult(key: String, value: T) {
     previousBackStackEntry?.savedStateHandle?.set(key, value)
 }
 
-fun PostCategory.toRequestString(): String? = when (this) {
-    PostCategory.NOTICE -> "notice"
-    PostCategory.QUIT_SMOKING_AIDS_REVIEWS -> "quit_smoking_aids_reviews"
-    PostCategory.SUCCESS_STORIES -> "success_stories"
-    PostCategory.GENERAL_DISCUSSION -> "general_discussion"
-    PostCategory.RESOLUTIONS -> "resolutions"
-    PostCategory.UNKNOWN -> null
-    PostCategory.ALL -> null
-}
 
-fun PostCategory.toStringKR(): String? = when (this) {
-    PostCategory.NOTICE -> "공지사항"
-    PostCategory.QUIT_SMOKING_AIDS_REVIEWS -> "금연 보조제 후기"
-    PostCategory.SUCCESS_STORIES -> "금연 성공 후기"
-    PostCategory.GENERAL_DISCUSSION -> "자유 게시판"
-    PostCategory.RESOLUTIONS -> "금연 다짐"
-    PostCategory.UNKNOWN -> null
-    PostCategory.ALL -> "커뮤니티 홈"
-}
 
-fun String.toPostCategory(): PostCategory = when (this) {
-    "notice", "공지사항" -> PostCategory.NOTICE
-    "quit_smoking_aids_reviews", "금연 보조제 후기" -> PostCategory.QUIT_SMOKING_AIDS_REVIEWS
-    "success_stories", "금연 성공 후기" -> PostCategory.SUCCESS_STORIES
-    "general_discussion", "자유 게시판" -> PostCategory.GENERAL_DISCUSSION
-    "resolutions", "금연 다짐" -> PostCategory.RESOLUTIONS
-    "all", "커뮤니티 홈" -> PostCategory.ALL
-    else -> PostCategory.UNKNOWN
-}
 
-fun PostWriteCategory.toRequestString(): String = when(this) {
-    PostWriteCategory.QUIT_SMOKING_AIDS_REVIEWS -> "quit_smoking_aids_reviews"
-    PostWriteCategory.SUCCESS_STORIES -> "success_stories"
-    PostWriteCategory.GENERAL_DISCUSSION -> "general_discussion"
-    PostWriteCategory.RESOLUTIONS -> "resolutions"
-}
-
-fun String.toPostWriteCategory() = when(this){
-    "자유 게시판" -> PostWriteCategory.GENERAL_DISCUSSION
-    "금연 성공 후기" -> PostWriteCategory.SUCCESS_STORIES
-    "금연 보조제 후기" -> PostWriteCategory.QUIT_SMOKING_AIDS_REVIEWS
-    "금연 다짐" -> PostWriteCategory.RESOLUTIONS
-    else -> throw IllegalStateException()
-}
-
-fun PostWriteCategory.toPostCategory(): PostCategory = when(this){
-    PostWriteCategory.GENERAL_DISCUSSION -> PostCategory.GENERAL_DISCUSSION
-    PostWriteCategory.SUCCESS_STORIES -> PostCategory.SUCCESS_STORIES
-    PostWriteCategory.QUIT_SMOKING_AIDS_REVIEWS -> PostCategory.QUIT_SMOKING_AIDS_REVIEWS
-    PostWriteCategory.RESOLUTIONS -> PostCategory.RESOLUTIONS
-    else -> PostCategory.UNKNOWN
-}
