+++ app/src/main/java/com/stopsmoke/kekkek/presentation/community/CommunityListAdapter.kt
@@ -10,7 +10,7 @@ import coil.load
 import com.stopsmoke.kekkek.R
 import com.stopsmoke.kekkek.databinding.ItemPostBinding
 import com.stopsmoke.kekkek.presentation.getRelativeTime
-import com.stopsmoke.kekkek.presentation.toStringKR
+import com.stopsmoke.kekkek.presentation.mapper.toStringKR
 import com.stopsmoke.kekkek.presentation.utils.diffutil.CommunityWritingItemDiffUtil
 
 class CommunityListAdapter :
