+++ app/src/main/java/com/stopsmoke/kekkek/presentation/post/edit/PostEditViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.viewModelScope
 import com.stopsmoke.kekkek.core.domain.model.Post
 import com.stopsmoke.kekkek.core.domain.model.PostEdit
 import com.stopsmoke.kekkek.core.domain.repository.PostRepository
-import com.stopsmoke.kekkek.presentation.toPostCategory
+import com.stopsmoke.kekkek.presentation.mapper.toPostCategory
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
