+++ app/src/main/java/com/stopsmoke/kekkek/presentation/search/post/SearchPostRecyclerViewAdapter.kt
@@ -9,7 +9,7 @@ import com.stopsmoke.kekkek.R
 import com.stopsmoke.kekkek.core.domain.model.Post
 import com.stopsmoke.kekkek.core.domain.model.ProfileImage
 import com.stopsmoke.kekkek.databinding.RecyclerviewSearchPostBinding
-import com.stopsmoke.kekkek.presentation.getResourceString
+import com.stopsmoke.kekkek.presentation.mapper.getResourceString
 import com.stopsmoke.kekkek.presentation.toResourceId
 import com.stopsmoke.kekkek.presentation.utils.diffutil.PostDiffUtil
