+++ app/src/main/java/com/stopsmoke/kekkek/presentation/userprofile/comment/adapter/UserProfileCommentListAdapter.kt
@@ -9,7 +9,7 @@ import com.stopsmoke.kekkek.R
 import com.stopsmoke.kekkek.core.domain.model.Comment
 import com.stopsmoke.kekkek.core.domain.model.CommentParent
 import com.stopsmoke.kekkek.databinding.RecyclerviewUserProfileCommentBinding
-import com.stopsmoke.kekkek.presentation.getResourceString
+import com.stopsmoke.kekkek.presentation.mapper.getResourceString
 import com.stopsmoke.kekkek.presentation.utils.diffutil.CommentDiffUtil
 
 class UserProfileCommentListAdapter(
